refactor: move getSimpleDiffChanges
